+++ packages/types/src/index.ts
@@ -1,2 +1,4 @@
 // exports will be generated automatically generated using: npm run generate:index
+export * from './eve';
+export * from './evm';
 export * from './system';