+++ packages/types/src/system/index.ts
@@ -1,3 +1,5 @@
+export type { default as IFetchOptions } from './IFetchOptions';
 export type { default as ILogger } from './ILogger';
 export type { default as TLogLevel } from './TLogLevel';
 export type { default as IUtilityOptions } from './IUtilityOptions';
+export type { default as TWithLastUpdatedAt } from './TWithLastUpdatedAt';